+++ lib/P2p/NetNodeCommon.h
@@ -21,8 +21,8 @@
 
 #pragma once
 
+#include <Global/CryptoNoteConfig.h>
 #include <P2p/P2pProtocolTypes.h>
-#include <../src/config/CryptoNoteConfig.h>
 #include <CryptoNote.h>
 
 namespace CryptoNote {
