+++ lib/Wallet/WalletGreen.h
@@ -21,6 +21,7 @@
 
 #include <queue>
 #include <unordered_map>
+#include <Global/CryptoNoteConfig.h>
 #include <Logging/LoggerRef.h>
 #include <System/Dispatcher.h>
 #include <System/Event.h>
@@ -29,7 +30,6 @@
 #include <Wallet/IFusionManager.h>
 #include <Wallet/IWallet.h>
 #include <Wallet/WalletIndices.h>
-#include <../src/config/CryptoNoteConfig.h>
 
 namespace CryptoNote {
